feat: propose config
